use chat history manager for title update
